Resolve incompatibilities for Linux 5.6 - 5.10
This is a single big commit that combines the multiple single commits
and the corresponding resolutions of merge conflicts, if any. The
combined commits are:

  * 029530e convert from "struct file_operations" to "struct proc_ops"
    to support kernel 5.6.0.
  * ced3f17 move private sha256_state definition out of header
  * 89189d6 linux kernel 5.8 compatibility
  * 8b3babe Fix compilation for Linux v5.10

The intent of combining all these in a single commit is to ease the
onboarding of newer driver versions. Assuming only easy to resolve merge
conflicts one only has to cherry pick one commit instead of four.