Allow compilation not booted kernels
This two line patch manifests a severe simplification of the workflow.

Now for i386, either KVER or KSRC can be specified as environment
variable and will be used later.

With this patch one can obtain a new kernel (Ubuntu Mainline, self
compiled, …), install it and test whether the driver can be compiled
without having to reboot. All one has to do is to set the environment
variable KVER.

Even simpler is the workflow if a clone of the Linux source code is
available. Then it is sufficient to set `KSRC=/path/to/linux/source`.
This makes it even simpler to check compilation, e.g. for older versions
of the kernel.