Fix compilation of not installed kernels
This commit fixes aac7c. The definitions of directories were neither
complete nor correct. With the present commit a compilation of Linux
5.12 succeeded.

In order to do so, the kernel has to be compiled. Doing

    make tinyconfig

was not sufficient, instead

    yes "" | make localmodconfig

had to be done. The compilation itself was done via `make dir-pkg` and
it was not tested whether other build targets work too.