Add DKMS helper scripts from xpadneo
The project xpadneo (https://github.com/atar-axis/xpadneo) has some
great DKMS helper scripts. It seems as if these are a better replacement
for our `deploy.sk` script.

This commit only adds the scripts almost unaltered. The only changes are
to adapt the import directory and to add a license header.