Revert "Use `--cvs-exclude` instead of hardcoding `.git/`"
This reverts commit f2bccb96ac8e54b63ef35a3c873075f21fb95030.

The option `--cvs-exclued` not only excluds CVS folders, but also
"core". The reason is given in
https://stackoverflow.com/a/9978573/8831116 . In a nutshell, CVS ignores
the folder "core" because it may contain core dumps.

Since we have a crucial folder "core", not syncing it breaks the DKMS
deployment, as reported in
https://github.com/cilynx/rtl88x2bu/issues/172.

Thus, the old behaviour has to be restored.