Fix one occassion of Werror=address
With GCC 12 gcc started to detect if comparisons to addresses are always
true. This could indicate a bug, as seen in the Linux kernel. For
instance, a simple overlooked ampersand could turn a useful comparison
to a tautological default, e.g. in `&my_struct->my_value`.

In this particular case the function guarded by the condition is empty
for Linux. It has implementations only for FreeBSD and Windows. This,
the entire block is removed.