Use tee to redirect content to root file
Previously, a plain `>` redirect was used. However, that does not work
if the user cannot write to the destination. Since a redirect is a
separate process, prepending `sudo` does not help neither, as it would
only affect `echo`.

The appropriate solution under GNU/Linux is to use `tee`. This is what
is done here.