Adapt type of change_beacon
Commit 66f85d57b7109baf8a7d5ee04049ac9412611d35 of the Linux Kernel
changed the type of one argument from cfg80211_beacon_data to
cfg80211_ap_settings. Inside the changed functions the beacon is
assigned to a local variable, so that subsequent usages do not have to
be adapted.

This commit follows the example. The type of the function argument is
changed and the beacon extracted.